Move all assignment components into "assigment" tab folder
