+++ src/containers/Review/assignment/AssignmentSectionRow.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react'
 import { Grid, Label, ModalProps } from 'semantic-ui-react'
-import ModalConfirmation from '../../components/Main/ModalConfirmation'
-import { useUserState } from '../../contexts/UserState'
-import { useLanguageProvider } from '../../contexts/Localisation'
-import { AssignmentOption } from '../../utils/data/assignmentOptions'
-import { useUnassignReviewAssignmentMutation } from '../../utils/generated/graphql'
-import useUpdateReviewAssignment from '../../utils/hooks/useUpdateReviewAssignment'
-import { AssignmentDetails, FullStructure } from '../../utils/types'
+import ModalConfirmation from '../../../components/Main/ModalConfirmation'
+import { useUserState } from '../../../contexts/UserState'
+import { useLanguageProvider } from '../../../contexts/Localisation'
+import { AssignmentOption } from '../../../utils/data/assignmentOptions'
+import { useUnassignReviewAssignmentMutation } from '../../../utils/generated/graphql'
+import useUpdateReviewAssignment from '../../../utils/hooks/useUpdateReviewAssignment'
+import { AssignmentDetails, FullStructure } from '../../../utils/types'
 import AssigneeDropdown from './AssigneeDropdown'
 import useGetAssignmentOptions from './useGetAssignmentOptions'
 import Reassignment from './Reassignment'
